create_embeddings/requirements.txt
requests
beautifulsoup4
openai
pinecone-client
site-map-parser
nltk